add `catalog::resolve_table_references` as a public utility
